% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estarfm_job.R
\name{estarfm_job}
\alias{estarfm_job}
\title{Title}
\usage{
estarfm_job(input_filenames, input_resolutions, input_dates, pred_dates,
  pred_filenames, pred_area, hightag, lowtag)
}
\arguments{
\item{input_filenames}{Gute Dateinamen}

\item{input_resolutions}{}

\item{input_dates}{}

\item{pred_dates}{}

\item{pred_filenames}{}

\item{pred_area}{}

\item{hightag}{}

\item{lowtag}{}
}
\value{
Something nice
}
\description{
Title
}
\examples{
Sorry, maybe later


  
}
