\name{ImageFusion-package}
\alias{ImageFusion-package}
\alias{ImageFusion}
\docType{package}
\title{
\packageTitle{ImageFusion}
}
\description{
\packageDescription{ImageFusion}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ImageFusion}
\packageIndices{ImageFusion}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ImageFusion}

Maintainer: \packageMaintainer{ImageFusion}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
