% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spstfm_job.R
\name{spstfm_job}
\alias{spstfm_job}
\title{Execute a single self-contained self-contained time-series imagefusion job using SPSTFM.}
\usage{
spstfm_job(
  input_filenames,
  input_resolutions,
  input_dates,
  pred_dates,
  pred_filenames,
  pred_area,
  winsize,
  date1,
  date3,
  n_cores,
  dict_size,
  n_training_samples,
  patch_size,
  patch_overlap,
  min_train_iter,
  max_train_iter,
  hightag,
  lowtag,
  MASKIMG_options,
  MASKRANGE_options,
  LOADDICT_options,
  SAVEDICT_options,
  REUSE_options,
  output_masks,
  use_nodata_value,
  use_parallelisation,
  random_sampling,
  verbose = T
)
}
\arguments{
\item{input_filenames}{A string vector containing the filenames of the input images}

\item{input_resolutions}{A string vector containing the resolution-tags (corresponding to the arguments \code{hightag} and \code{lowtag}, which are by default "high" and "low") of the input images.}

\item{input_dates}{An integer vector containing the dates of the input images.}

\item{pred_dates}{An integer vector  containing the dates for which images should be predicted.}

\item{pred_filenames}{A string vector containing the filenames for the predicted images. Must match \code{pred_dates} in length and order. Must include an extension relating to one of the \href{https://gdal.org/drivers/raster/index.html}{drivers supported by GDAL}, such as ".tif".}

\item{pred_area}{(Optional) An integer vector containing parameters in image coordinates for a bounding box which specifies the prediction area. The prediction will only be done in this area. (x_min, y_min, width, height). By default will use the entire area of the first input image.}

\item{date1}{(Optional) Set the date of the first input image pair. By default, will use the pair with the lowest date value.}

\item{date3}{(Optional) Set the date of the second input image pair. By default, will use the pair with the highest date value.}

\item{n_cores}{(Optional) Set the number of cores to use when using parallelisation. Note that parallelisation is currently not implemented for spstfm, and only listed here for consistency. Default is 1.}

\item{dict_size}{(Optional) Dictionary size, i. e. the number of atoms. Default: 256.}

\item{n_training_samples}{(Optional) The number of samples used for training (training data size). Default: 2000.}

\item{patch_size}{(Optional) Size of a patch in pixel. Default: 7.}

\item{patch_overlap}{(Optional) Overlap on each side of a patch in pixel. Default: 2.}

\item{min_train_iter}{(Optional) Minimum number of iterations during the training. Default: 10.}

\item{max_train_iter}{(Optional) Can be 0 if the minimum is also 0. Then no training will be done, even if there is no dictionary yet (it will still be initialized). Default: 20}

\item{hightag}{(Optional) A string which is used in \code{input_resolutions} to describe the high-resolution images. Default is "high".}

\item{lowtag}{(Optional) A string which is used in \code{input_resolutions} to describe the low-resolution images.  Default is "low".}

\item{MASKIMG_options}{(Optional) A string containing information for a mask image (8-bit, boolean, i. e. consists of 0 and 255). "For all input images the pixel values at the locations where the mask is 0 is replaced by the mean value." Example: \code{--mask-img=some_image.png}}

\item{MASKRANGE_options}{(Optional) Specify one or more intervals for valid values. Locations with invalid values will be masked out. Ranges should be given in the format '[<float>,<float>]', '(<float>,<float>)', '[<float>,<float>' or '<float>,<float>]'. There are a couple of options:' \itemize{
 \item{"--mask-valid-ranges"}{ Intervals which are marked as valid. Valid ranges can excluded from invalid ranges or vice versa, depending on the order of options.}
 \item{"--mask-invalid-ranges"}{ Intervals which are marked as invalid. Invalid intervals can be excluded from valid ranges or vice versa, depending on the order of options.}
 \item{"--mask-high-res-valid-ranges"}{ This is the same as --mask-valid-ranges, but is applied only for the high resolution images.}
 \item{"--mask-high-res-invalid-ranges"}{ This is the same as --mask-invalid-ranges, but is applied only for the high resolution images.}
 \item{"--mask-low-res-valid-ranges"}{ This is the same as --mask-valid-ranges, but is applied only for the low resolution images.}
 \item{"--mask-low-res-invalid-ranges"}{ This is the same as --mask-invalid-ranges, but is applied only for the low resolution images.}
}}

\item{LOADDICT_options}{(Optional) A string containing the path to a previously saved dictionary file, which is to be loaded (see \code{REUSE_OPTIONS}).}

\item{SAVEDICT_options}{(Optional) A string containing the path to a file to save the dictionary to after training.}

\item{REUSE_options}{(Optional) For a dictionary loaded from file, what to do with the existing dictionary: \itemize{
\item{"clear"} { clear an existing dictionary before training}
\item{"improve"} {improve an existing dictionary (default)}
\item{"use"} { use an existing dictionary without further training}
}}

\item{output_masks}{(Optional) Write mask images to disk? Default is "false".}

\item{use_nodata_value}{(Optional) Use the nodata value as invalid range for masking? Default is "true".}

\item{use_parallelisation}{(Optional) Use parallelisation when possible? Note that parallelisation is currently not implemented for spstfm, and only listed here for consistency. Default is "false".}

\item{random_sampling}{(Optional) Use random samples for training data instead of the samples with the most variance? Default is "false".}
}
\value{
Nothing. Output files are written to disk. The Geoinformation for the output images is adopted from the first input pair images.
}
\description{
A wrapper function for \code{execute_spstfm_job_cpp}. Intended to execute a single job, that is a number of predictions based on the same input pairs. It ensures that all of the arguments passed are of the correct type and creates sensible defaults.
}
\details{
Executes the SPSTFM algorithm to create a number of synthetic high-resolution images from two pairs of matching high- and low-resolution images.  Assumes that the input images already have matching size. For a detailed explanation how SPSTFM works there is the original paper and the thesis, which yielded this implementation. The latter explains also all available options and shows some test results. However, the default options should give good results.
}
\examples{
Sorry, maybe later
}
\references{
Huang, B., & Song, H. (2012). Spatiotemporal reflectance fusion via sparse representation. IEEE Transactions on Geoscience and Remote Sensing, 50(10), 3707-3716.
}
\seealso{
Other {fusion_algorithms}: 
\code{\link{estarfm_job}()},
\code{\link{fitfc_job}()},
\code{\link{starfm_job}()}
}
\author{
Christof Kaufmann (C++)

Johannes Mast (R)
}
\concept{{fusion_algorithms}}
