% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estarfm_job.R
\name{estarfm_job}
\alias{estarfm_job}
\title{Title}
\usage{
estarfm_job(input_filenames, input_resolutions, input_dates, pred_dates,
  pred_filenames, pred_area)
}
\value{
Nothinh
}
\description{
Title
}
\examples{





}
