% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFusion.R
\docType{package}
\name{ImageFusion}
\alias{ImageFusion}
\title{Imagefusion: A package for fusion of images}
\section{Outline}{

\code{ImageFusion} provides implementations of the following time-series fusion algorithms:
\itemize{
\item ESTARFM
\item FITFC
\item SPSTFM
\item STARFM
} 
More algorithms will be added over time.
}

\section{Implementation}{

The algorithms are implemented in C++ via 'GDAL', 'opencv' and 'Boost' and work from input images on disk without prior loading into R.
}

\author{
Christof Kaufmann (C++)

Johannes Mast (R)
}
