% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFusion.R
\docType{package}
\name{ImageFusion}
\alias{ImageFusion}
\title{Imagefusion: A package for fusion of images}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

