% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFusion_Task.R
\name{imagefusion_task}
\alias{imagefusion_task}
\title{imagefusion_task}
\usage{
imagefusion_task(..., filenames_high, filenames_low, dates_high, dates_low,
  dates_pred, singlepair_mode = "mixed", method = "starfm", verbose = T)
}
\arguments{
\item{...}{}

\item{filenames_high}{}

\item{filenames_low}{}

\item{dates_high}{}

\item{dates_low}{}

\item{dates_pred}{}

\item{singlepair_mode}{}

\item{method}{}

\item{verbose}{}
}
\value{
A ggplot overview of the tasks.
}
\description{
imagefusion_task
}
\examples{


}
